fix: ensuring secret generation is accurate (#578)
Description of changes:
When generating SetSDK, we were passing the memberName
to check if a field is a secret. When the secret field is nested, 
we need to pass memberFieldPath. That is what this change 
accomplishing

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.