add support for `reconcile-resources` flag (#580)
Description of changes:
https://github.com/aws-controllers-k8s/runtime/pull/176

This PR adds template changes to support the reconcile-resources flag which was recently added to the ACK runtime. This flag allows users to specify which resource kinds should be reconciled by a controller, providing more granular control over controller behavior.
When the flag is set, only the specified resource kinds will be reconciled by the controller. If unspecified, all resources will be reconciled (default behavior).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.