pass optype to setResourceReadMany from SetResource
Currently, if the resource output shape is a list, code-gen triggers the
readmany output generator.
When doing so, it does not pass the operationType, which causes it to
lose some of its functionality, which in this case was the ability to
ignore setting resources for a desired method.
With these changes, even sdk outputs that have a list as output type
will be able to configure Create/Update/Delete set config in
generator.yaml