fix: omitUnchangedFields except for status fields and required fields (#654)
Description of changes:
Currently omitUnchanged fields was applied to all spec fields.
This raises issues when certain required fields for update operations 
are in the spec. 

With this fix, we make sure to not omit required fields even if they are 
in spec

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.